+++ magpylib/_src/obj_classes/class_Collection.py
@@ -6,14 +6,13 @@ from magpylib._src.utility import (
     LIBRARY_SENSORS,
     LIBRARY_SOURCES)
 from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
-from magpylib._src.obj_classes.class_BaseRotate import BaseRotate
-from magpylib._src.obj_classes.class_BaseMove import BaseMove
+from magpylib._src.obj_classes.class_BaseTransform import BaseTransform
 from magpylib._src.fields.field_wrap_BH_level2 import getBH_level2
 from magpylib._src.default_utils import validate_style_keys
 from magpylib._src.exceptions import MagpylibBadUserInput
 
 
-class Collection(BaseDisplayRepr, BaseMove, BaseRotate):
+class BaseCollection(BaseDisplayRepr):
     """
     Group multiple objects in one Collection for common manipulation.
 
@@ -75,22 +74,6 @@ class Collection(BaseDisplayRepr, BaseMove, BaseRotate):
     >>> col - sphere
     >>> print(col.objects)
     [Loop(id=1879891543040)]
-
-    Manipulate all objects in a Collection directly using ``move`` and ``rotate`` methods
-
-    >>> import magpylib as magpy
-    >>> sphere = magpy.magnet.Sphere((1,2,3),1)
-    >>> loop = magpy.current.Loop(1,1)
-    >>> col = sphere + loop
-    >>> col.move((1,1,1))
-    >>> print(sphere.position)
-    [1. 1. 1.]
-
-    and compute the total magnetic field generated by the Collection.
-
-    >>> B = col.getB((1,2,3))
-    >>> print(B)
-    [-0.00372678  0.01820438  0.03423079]
 
     Consider three collections, a SourceCollection sCol a SensorCollection xCol and a
     MixedCollection mCol, all made up from the same objects.
@@ -470,3 +453,116 @@ class Collection(BaseDisplayRepr, BaseMove, BaseRotate):
         sources, sensors = self._validate_getBH_inputs(*objects)
 
         return getBH_level2(False, sources, sensors, sumup, squeeze)
+
+
+class Collection(BaseCollection, BaseTransform):
+    """
+    Group multiple objects in one Collection for common manipulation.
+
+    Operations applied to a Collection are sequentially applied to all objects in the Collection.
+    Collections do not allow duplicate objects (will be eliminated automatically).
+
+    Collections have the following dunders defined: __add__, __sub__, __iter__, __getitem__,
+    __repr__.
+
+    Depending on the input, Collection objects can be sources, observers or both. A Collection
+    with only source objects will become either a SourceCollection, one with only Sensors a
+    SensorCollection, and one with sources and sensors a MixedCollection
+    A SourceCollection functions like any SINGLE source. A SensorCollection functions like a
+    list of observer inputs. A MixedCollection will function as source or as observer.
+
+    Parameters
+    ----------
+    objects: sources, sensors, collections or arbitrary lists thereof
+        Ordered list of objects in the Collection.
+
+    Returns
+    -------
+    Collection object: Collection
+
+    Examples
+    --------
+
+    Create Collections for common manipulation. All objects added to a Collection
+    are stored in the ``objects`` attribute, and additionally in the ``sensors``
+    and ``sources`` attributes. These three return ordered sets (lists with
+    unique elements only)
+
+    >>> import magpylib as magpy
+    >>> sphere = magpy.magnet.Sphere((1,2,3),1)
+    >>> loop = magpy.current.Loop(1,1)
+    >>> dipole = magpy.misc.Dipole((1,2,3))
+    >>> col = magpy.Collection(sphere, loop, dipole)
+    >>> print(col.objects)
+    [Sphere(id=1879891544384), Loop(id=1879891543040), Dipole(id=1879892157152)]
+
+    Cycle directly through the Collection ``objects`` attribute
+
+    >>> for src in col:
+    >>>    print(src)
+    Sphere(id=1879891544384)
+    Loop(id=1879891543040)
+    Dipole(id=1879892157152)
+
+    and directly access objects from the Collection
+
+    >>> print(col[1])
+    Loop(id=1879891543040)
+
+    Add and subtract objects to form a Collection and to remove objects from a Collection.
+
+    >>> col = sphere + loop
+    >>> print(col.objects)
+    [Sphere(id=1879891544384), Loop(id=1879891543040)]
+    >>> col - sphere
+    >>> print(col.objects)
+    [Loop(id=1879891543040)]
+
+    Manipulate all objects in a Collection directly using ``move`` and ``rotate`` methods
+
+    >>> import magpylib as magpy
+    >>> sphere = magpy.magnet.Sphere((1,2,3),1)
+    >>> loop = magpy.current.Loop(1,1)
+    >>> col = sphere + loop
+    >>> col.move((1,1,1))
+    >>> print(sphere.position)
+    [1. 1. 1.]
+
+    and compute the total magnetic field generated by the Collection.
+
+    >>> B = col.getB((1,2,3))
+    >>> print(B)
+    [-0.00372678  0.01820438  0.03423079]
+
+    Consider three collections, a SourceCollection sCol a SensorCollection xCol and a
+    MixedCollection mCol, all made up from the same objects.
+
+    >>> import numpy as np
+    >>> import magpylib as magpy
+
+    >>> s1=magpy.magnet.Sphere((1,2,3), 1)
+    >>> s2=magpy.magnet.Cylinder((1,2,3), (1,1), (3,0,0))
+    >>> s3=magpy.magnet.Cuboid((1,2,3), (1,1,1), (6,0,0))
+
+    >>> x1=magpy.Sensor((1,0,3))
+    >>> x2=magpy.Sensor((4,0,3))
+    >>> x3=magpy.Sensor((7,0,3))
+
+    >>> sCol = magpy.Collection(s1, s2, s3)
+    >>> xCol = magpy.Collection(x1, x2, x3)
+    >>> mCol = magpy.Collection(sCol, xCol)
+
+    All the following lines will all give the same output
+
+    >>> magpy.getB([s1,s2,s3], [x1,x2,x3], sumup=True)
+    >>> magpy.getB(sCol, xCol)
+    >>> magpy.getB(mCol, mCol)
+    >>> sCol.getB(xCol)
+    >>> xCol.getB(sCol)
+    >>> sCol.getB(mCol)
+    >>> xCol.getB(mCol)
+    >>> mCol.getB()
+    """
+
+    def __init__(self, *objects):
+        super().__init__(*objects)
